refactor(frontend): 기존의 project list mockup data 제거, 신규로 대체
