--- frontend/src/constants/mockups/projects.js
@@ -1,75 +1,0 @@
-import defaultImage from "/public/images/default-project-thumbnail.png";
-import defaultUserImage from "/public/images/default-user-image.png";
-
-export const MOCKUP_PROJECT_LIST = [
-  {
-    id: "1",
-    subject: "해커톤",
-    title: "Uni-DTHON 파트너 하실 분 구합니다!",
-    startDate: "2022.11.05",
-    skills: ["Frontend", "React.js"],
-    author: {
-      name: "김성균",
-      photoUrl: defaultUserImage.src,
-    },
-    photoUrl: defaultImage.src,
-    approvalCount: 0,
-    capacity: 1,
-  },
-  {
-    id: "2",
-    subject: "해커톤",
-    title: "Uni-DTHON 파트너 하실 분 구합니다!",
-    startDate: "2022.11.05",
-    skills: ["Frontend", "React.js"],
-    author: {
-      name: "김성균",
-      photoUrl: defaultUserImage.src,
-    },
-    photoUrl: defaultImage.src,
-    approvalCount: 0,
-    capacity: 1,
-  },
-  {
-    id: "3",
-    subject: "해커톤",
-    title: "Uni-DTHON 파트너 하실 분 구합니다!",
-    startDate: "2022.11.05",
-    skills: ["Frontend", "React.js"],
-    author: {
-      name: "김성균",
-      photoUrl: defaultUserImage.src,
-    },
-    photoUrl: defaultImage.src,
-    approvalCount: 0,
-    capacity: 1,
-  },
-  {
-    id: "4",
-    subject: "해커톤",
-    title: "Uni-DTHON 파트너 하실 분 구합니다!",
-    startDate: "2022.11.05",
-    skills: ["Frontend", "React.js"],
-    author: {
-      name: "김성균",
-      photoUrl: defaultUserImage.src,
-    },
-    photoUrl: defaultImage.src,
-    approvalCount: 0,
-    capacity: 1,
-  },
-  {
-    id: "5",
-    subject: "해커톤",
-    title: "Uni-DTHON 파트너 하실 분 구합니다!",
-    startDate: "2022.11.05",
-    skills: ["Frontend", "React.js"],
-    author: {
-      name: "김성균",
-      photoUrl: defaultUserImage.src,
-    },
-    photoUrl: defaultImage.src,
-    approvalCount: 0,
-    capacity: 1,
-  },
-];
